requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP 요청
httpx==0.27.2
requests==2.32.3

# 웹 크롤링
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.26.1
webdriver-manager==4.0.2

# 데이터 처리
pydantic==2.10.3
python-dateutil==2.9.0

# 환경변수 관리
python-dotenv==1.0.1

# 로깅
loguru==0.7.2

# 메시지 브로커 (RabbitMQ)
pika==1.3.2

# ORM (PostgreSQL 직접 저장)
sqlalchemy==2.0.36
psycopg2-binary==2.9.10